renamed VerseCheckWrapperHooks.js to VerseCheckWrapper.js
